Refactor request model
Moves the address subdocument to its own file.
